scripts/docs-collator/requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
Deprecated==1.2.13
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
PyGithub==1.57
PyJWT==2.6.0
PyNaCl==1.5.0
requests==2.28.2
urllib3==1.26.14
wrapt==1.14.1
